requirements.txt
streamlit>=1.35.0
ultralytics>=8.2.0
Pillow>=10.0.0